requirements.txt
discord.py==2.3.2
pyautogui==0.9.54
